Move `Materializer` to `planning` package
