+++ pkg/streamingpromql/planning/materialize.go
@@ -1,27 +1,26 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package streamingpromql
+package planning
 
 import (
-	"github.com/grafana/mimir/pkg/streamingpromql/planning"
 	"github.com/grafana/mimir/pkg/streamingpromql/types"
 )
 
 // Materializer is responsible for converting query plan nodes to operators.
 // This type is not thread safe.
 type Materializer struct {
-	operatorFactories map[planning.Node]planning.OperatorFactory
-	operatorParams    *planning.OperatorParameters
+	operatorFactories map[Node]OperatorFactory
+	operatorParams    *OperatorParameters
 }
 
-func NewMaterializer(params *planning.OperatorParameters) *Materializer {
+func NewMaterializer(params *OperatorParameters) *Materializer {
 	return &Materializer{
-		operatorFactories: make(map[planning.Node]planning.OperatorFactory),
+		operatorFactories: make(map[Node]OperatorFactory),
 		operatorParams:    params,
 	}
 }
 
-func (m *Materializer) ConvertNodeToOperator(node planning.Node, timeRange types.QueryTimeRange) (types.Operator, error) {
+func (m *Materializer) ConvertNodeToOperator(node Node, timeRange types.QueryTimeRange) (types.Operator, error) {
 	// FIXME: we should check that we're not trying to get the same operator but with a different time range
 	if f, ok := m.operatorFactories[node]; ok {
 		return f.Produce()
